+++ internal-packages/run-engine/src/engine/eventBus.ts
@@ -1,6 +1,7 @@
 import { TaskRunExecutionStatus, TaskRunStatus } from "@trigger.dev/database";
 import { AuthenticatedEnvironment } from "../shared/index.js";
 import { FlushedRunMetadata, TaskRunError } from "@trigger.dev/core/v3";
+import { EventEmitter } from "events";
 
 export type EventBusEvents = {
   runAttemptStarted: [
@@ -178,3 +179,33 @@ export type EventBusEvents = {
 };
 
 export type EventBusEventArgs<T extends keyof EventBusEvents> = EventBusEvents[T];
+
+export type EventBus = EventEmitter<EventBusEvents>;
+
+/**
+ * Sends a notification that a run has changed and we need to fetch the latest run state.
+ * The worker will call `getRunExecutionData` via the API and act accordingly.
+ */
+export async function sendNotificationToWorker({
+  runId,
+  snapshot,
+  eventBus,
+}: {
+  runId: string;
+  snapshot: {
+    id: string;
+    executionStatus: TaskRunExecutionStatus;
+  };
+  eventBus: EventBus;
+}) {
+  eventBus.emit("workerNotification", {
+    time: new Date(),
+    run: {
+      id: runId,
+    },
+    snapshot: {
+      id: snapshot.id,
+      executionStatus: snapshot.executionStatus,
+    },
+  });
+}
