+++ internal-packages/run-engine/src/engine/index.ts
@@ -1,7 +1,6 @@
 import { createRedisClient, Redis } from "@internal/redis";
 import { Worker } from "@internal/redis-worker";
-import { Attributes, Span, SpanKind, trace, Tracer } from "@internal/tracing";
-import { assertExhaustive } from "@trigger.dev/core";
+import { startSpan, trace, Tracer } from "@internal/tracing";
 import { Logger } from "@trigger.dev/core/logger";
 import {
   CheckpointInput,
@@ -11,27 +10,19 @@ import {
   ExecutionResult,
   MachineResources,
   parsePacket,
-  RetryOptions,
   RunExecutionData,
   StartRunAttemptResult,
   TaskRunError,
-  TaskRunErrorCodes,
   TaskRunExecution,
   TaskRunExecutionResult,
-  TaskRunFailedExecutionResult,
-  TaskRunInternalError,
-  TaskRunSuccessfulExecutionResult,
   timeoutError,
 } from "@trigger.dev/core/v3";
 import {
   BatchId,
   CheckpointId,
-  getMaxDuration,
   parseNaturalLanguageDuration,
   QueueId,
   RunId,
-  sanitizeQueueName,
-  SnapshotId,
   WaitpointId,
 } from "@trigger.dev/core/v3/isomorphic";
 import {
@@ -39,96 +30,40 @@ import {
   Prisma,
   PrismaClient,
   PrismaClientOrTransaction,
-  RuntimeEnvironmentType,
   TaskRun,
   TaskRunExecutionSnapshot,
   TaskRunExecutionStatus,
-  TaskRunStatus,
   Waitpoint,
 } from "@trigger.dev/database";
 import { assertNever } from "assert-never";
 import { nanoid } from "nanoid";
 import { EventEmitter } from "node:events";
-import { z } from "zod";
 import { FairQueueSelectionStrategy } from "../run-queue/fairQueueSelectionStrategy.js";
 import { RunQueue } from "../run-queue/index.js";
 import { RunQueueFullKeyProducer } from "../run-queue/keyProducer.js";
 import { MinimalAuthenticatedEnvironment } from "../shared/index.js";
 import { MAX_TASK_RUN_ATTEMPTS } from "./consts.js";
-import { getRunWithBackgroundWorkerTasks } from "./db/worker.js";
-import { runStatusFromError } from "./errors.js";
-import { EventBusEvents } from "./eventBus.js";
-import { executionResultFromSnapshot, getLatestExecutionSnapshot } from "./executionSnapshots.js";
+import { EventBus, EventBusEvents, sendNotificationToWorker } from "./eventBus.js";
 import { RunLocker } from "./locking.js";
 import { getMachinePreset } from "./machinePresets.js";
-import { retryOutcomeFromCompletion } from "./retrying.js";
+import { ReleaseConcurrencyTokenBucketQueue } from "./releaseConcurrencyTokenBucketQueue.js";
 import {
   canReleaseConcurrency,
   isCheckpointable,
-  isDequeueableExecutionStatus,
   isExecuting,
-  isFinalRunStatus,
   isPendingExecuting,
 } from "./statuses.js";
-import { HeartbeatTimeouts, RunEngineOptions, TriggerParams } from "./types.js";
-import { ReleaseConcurrencyTokenBucketQueue } from "./releaseConcurrencyTokenBucketQueue.js";
-
-const workerCatalog = {
-  finishWaitpoint: {
-    schema: z.object({
-      waitpointId: z.string(),
-      error: z.string().optional(),
-    }),
-    visibilityTimeoutMs: 5000,
-  },
-  heartbeatSnapshot: {
-    schema: z.object({
-      runId: z.string(),
-      snapshotId: z.string(),
-    }),
-    visibilityTimeoutMs: 5000,
-  },
-  expireRun: {
-    schema: z.object({
-      runId: z.string(),
-    }),
-    visibilityTimeoutMs: 5000,
-  },
-  cancelRun: {
-    schema: z.object({
-      runId: z.string(),
-      completedAt: z.coerce.date(),
-      reason: z.string().optional(),
-    }),
-    visibilityTimeoutMs: 5000,
-  },
-  queueRunsWaitingForWorker: {
-    schema: z.object({
-      backgroundWorkerId: z.string(),
-    }),
-    visibilityTimeoutMs: 5000,
-  },
-  tryCompleteBatch: {
-    schema: z.object({
-      batchId: z.string(),
-    }),
-    visibilityTimeoutMs: 10_000,
-  },
-  continueRunIfUnblocked: {
-    schema: z.object({
-      runId: z.string(),
-    }),
-    visibilityTimeoutMs: 10_000,
-  },
-  enqueueDelayedRun: {
-    schema: z.object({
-      runId: z.string(),
-    }),
-    visibilityTimeoutMs: 10_000,
-  },
-};
-
-type EngineWorker = Worker<typeof workerCatalog>;
+import { BatchSystem } from "./systems/batchSystem.js";
+import { DequeueSystem } from "./systems/dequeueSystem.js";
+import {
+  executionResultFromSnapshot,
+  ExecutionSnapshotSystem,
+  getLatestExecutionSnapshot,
+} from "./systems/executionSnapshotSystem.js";
+import { RunAttemptSystem } from "./systems/runAttemptSystem.js";
+import { WaitpointSystem } from "./systems/waitpointSystem.js";
+import { EngineWorker, HeartbeatTimeouts, RunEngineOptions, TriggerParams } from "./types.js";
+import { workerCatalog } from "./workerCatalog.js";
 
 export class RunEngine {
   private runLockRedis: Redis;
@@ -144,7 +79,12 @@ export class RunEngine {
     projectId: string;
     envId: string;
   }>;
-  eventBus = new EventEmitter<EventBusEvents>();
+  eventBus: EventBus = new EventEmitter<EventBusEvents>();
+  executionSnapshotSystem: ExecutionSnapshotSystem;
+  runAttemptSystem: RunAttemptSystem;
+  dequeueSystem: DequeueSystem;
+  waitpointSystem: WaitpointSystem;
+  batchSystem: BatchSystem;
 
   constructor(private readonly options: RunEngineOptions) {
     this.prisma = options.prisma;
@@ -211,7 +151,7 @@ export class RunEngine {
           await this.#expireRun({ runId: payload.runId });
         },
         cancelRun: async ({ payload }) => {
-          await this.cancelRun({
+          await this.runAttemptSystem.cancelRun({
             runId: payload.runId,
             completedAt: payload.completedAt,
             reason: payload.reason,
@@ -221,7 +161,7 @@ export class RunEngine {
           await this.#queueRunsWaitingForWorker({ backgroundWorkerId: payload.backgroundWorkerId });
         },
         tryCompleteBatch: async ({ payload }) => {
-          await this.#tryCompleteBatch({ batchId: payload.batchId });
+          await this.batchSystem.performCompleteBatch({ batchId: payload.batchId });
         },
         continueRunIfUnblocked: async ({ payload }) => {
           await this.#continueRunIfUnblocked({
@@ -291,6 +231,54 @@ export class RunEngine {
       },
       tracer: this.tracer,
     });
+
+    this.executionSnapshotSystem = new ExecutionSnapshotSystem({
+      worker: this.worker,
+      eventBus: this.eventBus,
+      heartbeatTimeouts: this.heartbeatTimeouts,
+      prisma: this.prisma,
+      logger: this.logger,
+      tracer: this.tracer,
+    });
+
+    this.waitpointSystem = new WaitpointSystem({
+      prisma: this.prisma,
+      worker: this.worker,
+      eventBus: this.eventBus,
+      logger: this.logger,
+      tracer: this.tracer,
+    });
+
+    this.batchSystem = new BatchSystem({
+      prisma: this.prisma,
+      logger: this.logger,
+      tracer: this.tracer,
+      worker: this.worker,
+    });
+
+    this.runAttemptSystem = new RunAttemptSystem({
+      prisma: this.prisma,
+      logger: this.logger,
+      tracer: this.tracer,
+      runLock: this.runLock,
+      eventBus: this.eventBus,
+      runQueue: this.runQueue,
+      worker: this.worker,
+      executionSnapshotSystem: this.executionSnapshotSystem,
+      batchSystem: this.batchSystem,
+      waitpointSystem: this.waitpointSystem,
+    });
+
+    this.dequeueSystem = new DequeueSystem({
+      prisma: this.prisma,
+      queue: this.runQueue,
+      runLock: this.runLock,
+      logger: this.logger,
+      machines: this.options.machines,
+      tracer: this.tracer,
+      executionSnapshotSystem: this.executionSnapshotSystem,
+      runAttemptSystem: this.runAttemptSystem,
+    });
   }
 
   //MARK: - Run functions
@@ -346,14 +334,9 @@ export class RunEngine {
   ): Promise<TaskRun> {
     const prisma = tx ?? this.prisma;
 
-    return this.#trace(
+    return startSpan(
+      this.tracer,
       "trigger",
-      {
-        friendlyId,
-        environmentId: environment.id,
-        projectId: environment.project.id,
-        taskIdentifier,
-      },
       async (span) => {
         const status = delayUntil ? "DELAYED" : "PENDING";
 
@@ -596,6 +579,14 @@ export class RunEngine {
         });
 
         return taskRun;
+      },
+      {
+        attributes: {
+          friendlyId,
+          environmentId: environment.id,
+          projectId: environment.project.id,
+          taskIdentifier,
+        },
       }
     );
   }
@@ -625,437 +616,15 @@ export class RunEngine {
     runnerId?: string;
     tx?: PrismaClientOrTransaction;
   }): Promise<DequeuedMessage[]> {
-    const prisma = tx ?? this.prisma;
-    return this.#trace("dequeueFromMasterQueue", { consumerId, masterQueue }, async (span) => {
-      //gets multiple runs from the queue
-      const messages = await this.runQueue.dequeueMessageFromMasterQueue(
-        consumerId,
-        masterQueue,
-        maxRunCount
-      );
-      if (messages.length === 0) {
-        return [];
-      }
-
-      //we can't send more than the max resources
-      const consumedResources: MachineResources = {
-        cpu: 0,
-        memory: 0,
-      };
-
-      const dequeuedRuns: DequeuedMessage[] = [];
-
-      for (const message of messages) {
-        const orgId = message.message.orgId;
-        const runId = message.messageId;
-
-        span.setAttribute("runId", runId);
-
-        //lock the run so nothing else can modify it
-        try {
-          const dequeuedRun = await this.runLock.lock([runId], 5000, async (signal) => {
-            const snapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-            if (!isDequeueableExecutionStatus(snapshot.executionStatus)) {
-              //create a failed snapshot
-              await this.#createExecutionSnapshot(prisma, {
-                run: {
-                  id: snapshot.runId,
-                  status: snapshot.runStatus,
-                },
-                snapshot: {
-                  executionStatus: snapshot.executionStatus,
-                  description:
-                    "Tried to dequeue a run that is not in a valid state to be dequeued.",
-                },
-                environmentId: snapshot.environmentId,
-                environmentType: snapshot.environmentType,
-                checkpointId: snapshot.checkpointId ?? undefined,
-                completedWaitpoints: snapshot.completedWaitpoints,
-                error: `Tried to dequeue a run that is not in a valid state to be dequeued.`,
-                workerId,
-                runnerId,
-              });
-
-              //todo is there a way to recover this, so the run can be retried?
-              //for example should we update the status to a dequeuable status and nack it?
-              //then at least it has a chance of succeeding and we have the error log above
-              await this.#systemFailure({
-                runId,
-                error: {
-                  type: "INTERNAL_ERROR",
-                  code: "TASK_DEQUEUED_INVALID_STATE",
-                  message: `Task was in the ${snapshot.executionStatus} state when it was dequeued for execution.`,
-                },
-                tx: prisma,
-              });
-              this.logger.error(
-                `RunEngine.dequeueFromMasterQueue(): Run is not in a valid state to be dequeued: ${runId}\n ${snapshot.id}:${snapshot.executionStatus}`
-              );
-              return null;
-            }
-
-            const result = await getRunWithBackgroundWorkerTasks(prisma, runId, backgroundWorkerId);
-
-            if (!result.success) {
-              switch (result.code) {
-                case "NO_RUN": {
-                  //this should not happen, the run is unrecoverable so we'll ack it
-                  this.logger.error("RunEngine.dequeueFromMasterQueue(): No run found", {
-                    runId,
-                    latestSnapshot: snapshot.id,
-                  });
-                  await this.runQueue.acknowledgeMessage(orgId, runId);
-                  return null;
-                }
-                case "NO_WORKER":
-                case "TASK_NEVER_REGISTERED":
-                case "TASK_NOT_IN_LATEST": {
-                  this.logger.warn(`RunEngine.dequeueFromMasterQueue(): ${result.code}`, {
-                    runId,
-                    latestSnapshot: snapshot.id,
-                    result,
-                  });
-
-                  //not deployed yet, so we'll wait for the deploy
-                  await this.#waitingForDeploy({
-                    orgId,
-                    runId,
-                    reason: result.message,
-                    tx: prisma,
-                  });
-                  return null;
-                }
-                case "BACKGROUND_WORKER_MISMATCH": {
-                  this.logger.warn(
-                    "RunEngine.dequeueFromMasterQueue(): Background worker mismatch",
-                    {
-                      runId,
-                      latestSnapshot: snapshot.id,
-                      result,
-                    }
-                  );
-
-                  //worker mismatch so put it back in the queue
-                  await this.runQueue.nackMessage({ orgId, messageId: runId });
-
-                  return null;
-                }
-                default: {
-                  assertExhaustive(result);
-                }
-              }
-            }
-
-            //check for a valid deployment if it's not a development environment
-            if (result.run.runtimeEnvironment.type !== "DEVELOPMENT") {
-              if (!result.deployment || !result.deployment.imageReference) {
-                this.logger.warn("RunEngine.dequeueFromMasterQueue(): No deployment found", {
-                  runId,
-                  latestSnapshot: snapshot.id,
-                  result,
-                });
-                //not deployed yet, so we'll wait for the deploy
-                await this.#waitingForDeploy({
-                  orgId,
-                  runId,
-                  reason: "No deployment or deployment image reference found for deployed run",
-                  tx: prisma,
-                });
-
-                return null;
-              }
-            }
-
-            const machinePreset = getMachinePreset({
-              machines: this.options.machines.machines,
-              defaultMachine: this.options.machines.defaultMachine,
-              config: result.task.machineConfig ?? {},
-              run: result.run,
-            });
-
-            //increment the consumed resources
-            consumedResources.cpu += machinePreset.cpu;
-            consumedResources.memory += machinePreset.memory;
-
-            //are we under the limit?
-            if (maxResources) {
-              if (
-                consumedResources.cpu > maxResources.cpu ||
-                consumedResources.memory > maxResources.memory
-              ) {
-                this.logger.debug(
-                  "RunEngine.dequeueFromMasterQueue(): Consumed resources over limit, nacking",
-                  {
-                    runId,
-                    consumedResources,
-                    maxResources,
-                  }
-                );
-
-                //put it back in the queue where it was
-                await this.runQueue.nackMessage({
-                  orgId,
-                  messageId: runId,
-                  incrementAttemptCount: false,
-                  retryAt: result.run.createdAt.getTime() - result.run.priorityMs,
-                });
-                return null;
-              }
-            }
-
-            // Check max attempts that can optionally be set when triggering a run
-            let maxAttempts: number | null | undefined = result.run.maxAttempts;
-
-            // If it's not set, we'll grab it from the task's retry config
-            if (!maxAttempts) {
-              const retryConfig = result.task.retryConfig;
-
-              this.logger.debug(
-                "RunEngine.dequeueFromMasterQueue(): maxAttempts not set, using task's retry config",
-                {
-                  runId,
-                  task: result.task.id,
-                  rawRetryConfig: retryConfig,
-                }
-              );
-
-              const parsedConfig = RetryOptions.nullable().safeParse(retryConfig);
-
-              if (!parsedConfig.success) {
-                this.logger.error("RunEngine.dequeueFromMasterQueue(): Invalid retry config", {
-                  runId,
-                  task: result.task.id,
-                  rawRetryConfig: retryConfig,
-                });
-
-                await this.#systemFailure({
-                  runId,
-                  error: {
-                    type: "INTERNAL_ERROR",
-                    code: "TASK_DEQUEUED_INVALID_RETRY_CONFIG",
-                    message: `Invalid retry config: ${retryConfig}`,
-                  },
-                  tx: prisma,
-                });
-
-                return null;
-              }
-
-              if (!parsedConfig.data) {
-                this.logger.error("RunEngine.dequeueFromMasterQueue(): No retry config", {
-                  runId,
-                  task: result.task.id,
-                  rawRetryConfig: retryConfig,
-                });
-
-                await this.#systemFailure({
-                  runId,
-                  error: {
-                    type: "INTERNAL_ERROR",
-                    code: "TASK_DEQUEUED_NO_RETRY_CONFIG",
-                    message: `No retry config found`,
-                  },
-                  tx: prisma,
-                });
-
-                return null;
-              }
-
-              maxAttempts = parsedConfig.data.maxAttempts;
-            }
-
-            //update the run
-            const lockedTaskRun = await prisma.taskRun.update({
-              where: {
-                id: runId,
-              },
-              data: {
-                lockedAt: new Date(),
-                lockedById: result.task.id,
-                lockedToVersionId: result.worker.id,
-                startedAt: result.run.startedAt ?? new Date(),
-                baseCostInCents: this.options.machines.baseCostInCents,
-                machinePreset: machinePreset.name,
-                taskVersion: result.worker.version,
-                sdkVersion: result.worker.sdkVersion,
-                cliVersion: result.worker.cliVersion,
-                maxDurationInSeconds: getMaxDuration(
-                  result.run.maxDurationInSeconds,
-                  result.task.maxDurationInSeconds
-                ),
-                maxAttempts: maxAttempts ?? undefined,
-              },
-              include: {
-                runtimeEnvironment: true,
-                tags: true,
-              },
-            });
-
-            if (!lockedTaskRun) {
-              this.logger.error("RunEngine.dequeueFromMasterQueue(): Failed to lock task run", {
-                taskRun: result.run.id,
-                taskIdentifier: result.run.taskIdentifier,
-                deployment: result.deployment?.id,
-                worker: result.worker.id,
-                task: result.task.id,
-                runId,
-              });
-
-              await this.runQueue.acknowledgeMessage(orgId, runId);
-              return null;
-            }
-
-            const queue = await prisma.taskQueue.findUnique({
-              where: {
-                runtimeEnvironmentId_name: {
-                  runtimeEnvironmentId: lockedTaskRun.runtimeEnvironmentId,
-                  name: sanitizeQueueName(lockedTaskRun.queue),
-                },
-              },
-            });
-
-            if (!queue) {
-              this.logger.debug(
-                "RunEngine.dequeueFromMasterQueue(): queue not found, so nacking message",
-                {
-                  queueMessage: message,
-                  taskRunQueue: lockedTaskRun.queue,
-                  runtimeEnvironmentId: lockedTaskRun.runtimeEnvironmentId,
-                }
-              );
-
-              //will auto-retry
-              const gotRequeued = await this.runQueue.nackMessage({ orgId, messageId: runId });
-              if (!gotRequeued) {
-                await this.#systemFailure({
-                  runId,
-                  error: {
-                    type: "INTERNAL_ERROR",
-                    code: "TASK_DEQUEUED_QUEUE_NOT_FOUND",
-                    message: `Tried to dequeue the run but the queue doesn't exist: ${lockedTaskRun.queue}`,
-                  },
-                  tx: prisma,
-                });
-              }
-
-              return null;
-            }
-
-            const currentAttemptNumber = lockedTaskRun.attemptNumber ?? 0;
-            const nextAttemptNumber = currentAttemptNumber + 1;
-
-            const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-              run: {
-                id: runId,
-                status: snapshot.runStatus,
-                attemptNumber: lockedTaskRun.attemptNumber,
-              },
-              snapshot: {
-                executionStatus: "PENDING_EXECUTING",
-                description: "Run was dequeued for execution",
-              },
-              environmentId: snapshot.environmentId,
-              environmentType: snapshot.environmentType,
-              checkpointId: snapshot.checkpointId ?? undefined,
-              completedWaitpoints: snapshot.completedWaitpoints,
-              workerId,
-              runnerId,
-            });
-
-            return {
-              version: "1" as const,
-              dequeuedAt: new Date(),
-              snapshot: {
-                id: newSnapshot.id,
-                friendlyId: newSnapshot.friendlyId,
-                executionStatus: newSnapshot.executionStatus,
-                description: newSnapshot.description,
-              },
-              image: result.deployment?.imageReference ?? undefined,
-              checkpoint: newSnapshot.checkpoint ?? undefined,
-              completedWaitpoints: snapshot.completedWaitpoints,
-              backgroundWorker: {
-                id: result.worker.id,
-                friendlyId: result.worker.friendlyId,
-                version: result.worker.version,
-              },
-              deployment: {
-                id: result.deployment?.id,
-                friendlyId: result.deployment?.friendlyId,
-              },
-              run: {
-                id: lockedTaskRun.id,
-                friendlyId: lockedTaskRun.friendlyId,
-                isTest: lockedTaskRun.isTest,
-                machine: machinePreset,
-                attemptNumber: nextAttemptNumber,
-                masterQueue: lockedTaskRun.masterQueue,
-                traceContext: lockedTaskRun.traceContext as Record<string, unknown>,
-              },
-              environment: {
-                id: lockedTaskRun.runtimeEnvironment.id,
-                type: lockedTaskRun.runtimeEnvironment.type,
-              },
-              organization: {
-                id: orgId,
-              },
-              project: {
-                id: lockedTaskRun.projectId,
-              },
-            } satisfies DequeuedMessage;
-          });
-
-          if (dequeuedRun !== null) {
-            dequeuedRuns.push(dequeuedRun);
-          }
-        } catch (error) {
-          this.logger.error(
-            "RunEngine.dequeueFromMasterQueue(): Thrown error while preparing run to be run",
-            {
-              error,
-              runId,
-            }
-          );
-
-          const run = await prisma.taskRun.findFirst({
-            where: { id: runId },
-            include: {
-              runtimeEnvironment: true,
-            },
-          });
-
-          if (!run) {
-            //this isn't ideal because we're not creating a snapshot… but we can't do much else
-            this.logger.error(
-              "RunEngine.dequeueFromMasterQueue(): Thrown error, then run not found. Nacking.",
-              {
-                runId,
-                orgId,
-              }
-            );
-            await this.runQueue.nackMessage({ orgId, messageId: runId });
-            continue;
-          }
-
-          //this is an unknown error, we'll reattempt (with auto-backoff and eventually DLQ)
-          const gotRequeued = await this.#tryNackAndRequeue({
-            run,
-            environment: run.runtimeEnvironment,
-            orgId,
-            error: {
-              type: "INTERNAL_ERROR",
-              code: "TASK_RUN_DEQUEUED_MAX_RETRIES",
-              message: `We tried to dequeue the run the maximum number of times but it wouldn't start executing`,
-            },
-            tx: prisma,
-          });
-          //we don't need this, but it makes it clear we're in a loop here
-          continue;
-        }
-      }
-
-      return dequeuedRuns;
+    return this.dequeueSystem.dequeueFromMasterQueue({
+      consumerId,
+      masterQueue,
+      maxRunCount,
+      maxResources,
+      backgroundWorkerId,
+      workerId,
+      runnerId,
+      tx,
     });
   }
 
@@ -1136,266 +705,273 @@ export class RunEngine {
   }): Promise<StartRunAttemptResult> {
     const prisma = tx ?? this.prisma;
 
-    return this.#trace("startRunAttempt", { runId, snapshotId }, async (span) => {
-      return this.runLock.lock([runId], 5000, async (signal) => {
-        const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
+    return startSpan(
+      this.tracer,
+      "startRunAttempt",
+      async (span) => {
+        return this.runLock.lock([runId], 5000, async () => {
+          const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
 
-        if (latestSnapshot.id !== snapshotId) {
-          //if there is a big delay between the snapshot and the attempt, the snapshot might have changed
-          //we just want to log because elsewhere it should have been put back into a state where it can be attempted
-          this.logger.warn(
-            "RunEngine.createRunAttempt(): snapshot has changed since the attempt was created, ignoring."
-          );
-          throw new ServiceValidationError("Snapshot changed", 409);
-        }
+          if (latestSnapshot.id !== snapshotId) {
+            //if there is a big delay between the snapshot and the attempt, the snapshot might have changed
+            //we just want to log because elsewhere it should have been put back into a state where it can be attempted
+            this.logger.warn(
+              "RunEngine.createRunAttempt(): snapshot has changed since the attempt was created, ignoring."
+            );
+            throw new ServiceValidationError("Snapshot changed", 409);
+          }
 
-        const environment = await this.#getAuthenticatedEnvironmentFromRun(runId, prisma);
-        if (!environment) {
-          throw new ServiceValidationError("Environment not found", 404);
-        }
+          const environment = await this.#getAuthenticatedEnvironmentFromRun(runId, prisma);
+          if (!environment) {
+            throw new ServiceValidationError("Environment not found", 404);
+          }
 
-        const taskRun = await prisma.taskRun.findFirst({
-          where: {
-            id: runId,
-          },
-          include: {
-            tags: true,
-            lockedBy: {
-              include: {
-                worker: {
-                  select: {
-                    id: true,
-                    version: true,
-                    sdkVersion: true,
-                    cliVersion: true,
-                    supportsLazyAttempts: true,
+          const taskRun = await prisma.taskRun.findFirst({
+            where: {
+              id: runId,
+            },
+            include: {
+              tags: true,
+              lockedBy: {
+                include: {
+                  worker: {
+                    select: {
+                      id: true,
+                      version: true,
+                      sdkVersion: true,
+                      cliVersion: true,
+                      supportsLazyAttempts: true,
+                    },
                   },
                 },
               },
-            },
-            batchItems: {
-              include: {
-                batchTaskRun: true,
-              },
-            },
-          },
-        });
-
-        this.logger.debug("Creating a task run attempt", { taskRun });
-
-        if (!taskRun) {
-          throw new ServiceValidationError("Task run not found", 404);
-        }
-
-        span.setAttribute("projectId", taskRun.projectId);
-        span.setAttribute("environmentId", taskRun.runtimeEnvironmentId);
-        span.setAttribute("taskRunId", taskRun.id);
-        span.setAttribute("taskRunFriendlyId", taskRun.friendlyId);
-
-        if (taskRun.status === "CANCELED") {
-          throw new ServiceValidationError("Task run is cancelled", 400);
-        }
-
-        if (!taskRun.lockedBy) {
-          throw new ServiceValidationError("Task run is not locked", 400);
-        }
-
-        const queue = await prisma.taskQueue.findUnique({
-          where: {
-            runtimeEnvironmentId_name: {
-              runtimeEnvironmentId: environment.id,
-              name: taskRun.queue,
-            },
-          },
-        });
-
-        if (!queue) {
-          throw new ServiceValidationError("Queue not found", 404);
-        }
-
-        //increment the attempt number (start at 1)
-        const nextAttemptNumber = (taskRun.attemptNumber ?? 0) + 1;
-
-        if (nextAttemptNumber > MAX_TASK_RUN_ATTEMPTS) {
-          await this.#attemptFailed({
-            runId: taskRun.id,
-            snapshotId,
-            completion: {
-              ok: false,
-              id: taskRun.id,
-              error: {
-                type: "INTERNAL_ERROR",
-                code: "TASK_RUN_CRASHED",
-                message: "Max attempts reached.",
-              },
-            },
-            tx: prisma,
-          });
-          throw new ServiceValidationError("Max attempts reached", 400);
-        }
-
-        this.eventBus.emit("runAttemptStarted", {
-          time: new Date(),
-          run: {
-            id: taskRun.id,
-            attemptNumber: nextAttemptNumber,
-            baseCostInCents: taskRun.baseCostInCents,
-          },
-          organization: {
-            id: environment.organization.id,
-          },
-        });
-
-        const result = await $transaction(
-          prisma,
-          async (tx) => {
-            const run = await tx.taskRun.update({
-              where: {
-                id: taskRun.id,
-              },
-              data: {
-                status: "EXECUTING",
-                attemptNumber: nextAttemptNumber,
-                executedAt: taskRun.attemptNumber === null ? new Date() : undefined,
-              },
-              include: {
-                tags: true,
-                lockedBy: {
-                  include: { worker: true },
+              batchItems: {
+                include: {
+                  batchTaskRun: true,
                 },
               },
-            });
-
-            const newSnapshot = await this.#createExecutionSnapshot(tx, {
-              run,
-              snapshot: {
-                executionStatus: "EXECUTING",
-                description: `Attempt created, starting execution${
-                  isWarmStart ? " (warm start)" : ""
-                }`,
-              },
-              environmentId: latestSnapshot.environmentId,
-              environmentType: latestSnapshot.environmentType,
-              workerId,
-              runnerId,
-            });
-
-            if (taskRun.ttl) {
-              //don't expire the run, it's going to execute
-              await this.worker.ack(`expireRun:${taskRun.id}`);
-            }
-
-            return { run, snapshot: newSnapshot };
-          },
-          (error) => {
-            this.logger.error("RunEngine.createRunAttempt(): prisma.$transaction error", {
-              code: error.code,
-              meta: error.meta,
-              stack: error.stack,
-              message: error.message,
-              name: error.name,
-            });
-            throw new ServiceValidationError(
-              "Failed to update task run and execution snapshot",
-              500
-            );
-          }
-        );
-
-        if (!result) {
-          this.logger.error("RunEngine.createRunAttempt(): failed to create task run attempt", {
-            runId: taskRun.id,
-            nextAttemptNumber,
+            },
           });
-          throw new ServiceValidationError("Failed to create task run attempt", 500);
-        }
 
-        const { run, snapshot } = result;
+          this.logger.debug("Creating a task run attempt", { taskRun });
 
-        const machinePreset = getMachinePreset({
-          machines: this.options.machines.machines,
-          defaultMachine: this.options.machines.defaultMachine,
-          config: taskRun.lockedBy.machineConfig ?? {},
-          run: taskRun,
+          if (!taskRun) {
+            throw new ServiceValidationError("Task run not found", 404);
+          }
+
+          span.setAttribute("projectId", taskRun.projectId);
+          span.setAttribute("environmentId", taskRun.runtimeEnvironmentId);
+          span.setAttribute("taskRunId", taskRun.id);
+          span.setAttribute("taskRunFriendlyId", taskRun.friendlyId);
+
+          if (taskRun.status === "CANCELED") {
+            throw new ServiceValidationError("Task run is cancelled", 400);
+          }
+
+          if (!taskRun.lockedBy) {
+            throw new ServiceValidationError("Task run is not locked", 400);
+          }
+
+          const queue = await prisma.taskQueue.findUnique({
+            where: {
+              runtimeEnvironmentId_name: {
+                runtimeEnvironmentId: environment.id,
+                name: taskRun.queue,
+              },
+            },
+          });
+
+          if (!queue) {
+            throw new ServiceValidationError("Queue not found", 404);
+          }
+
+          //increment the attempt number (start at 1)
+          const nextAttemptNumber = (taskRun.attemptNumber ?? 0) + 1;
+
+          if (nextAttemptNumber > MAX_TASK_RUN_ATTEMPTS) {
+            await this.runAttemptSystem.attemptFailed({
+              runId: taskRun.id,
+              snapshotId,
+              completion: {
+                ok: false,
+                id: taskRun.id,
+                error: {
+                  type: "INTERNAL_ERROR",
+                  code: "TASK_RUN_CRASHED",
+                  message: "Max attempts reached.",
+                },
+              },
+              tx: prisma,
+            });
+            throw new ServiceValidationError("Max attempts reached", 400);
+          }
+
+          this.eventBus.emit("runAttemptStarted", {
+            time: new Date(),
+            run: {
+              id: taskRun.id,
+              attemptNumber: nextAttemptNumber,
+              baseCostInCents: taskRun.baseCostInCents,
+            },
+            organization: {
+              id: environment.organization.id,
+            },
+          });
+
+          const result = await $transaction(
+            prisma,
+            async (tx) => {
+              const run = await tx.taskRun.update({
+                where: {
+                  id: taskRun.id,
+                },
+                data: {
+                  status: "EXECUTING",
+                  attemptNumber: nextAttemptNumber,
+                  executedAt: taskRun.attemptNumber === null ? new Date() : undefined,
+                },
+                include: {
+                  tags: true,
+                  lockedBy: {
+                    include: { worker: true },
+                  },
+                },
+              });
+
+              const newSnapshot = await this.executionSnapshotSystem.createExecutionSnapshot(tx, {
+                run,
+                snapshot: {
+                  executionStatus: "EXECUTING",
+                  description: `Attempt created, starting execution${
+                    isWarmStart ? " (warm start)" : ""
+                  }`,
+                },
+                environmentId: latestSnapshot.environmentId,
+                environmentType: latestSnapshot.environmentType,
+                workerId,
+                runnerId,
+              });
+
+              if (taskRun.ttl) {
+                //don't expire the run, it's going to execute
+                await this.worker.ack(`expireRun:${taskRun.id}`);
+              }
+
+              return { run, snapshot: newSnapshot };
+            },
+            (error) => {
+              this.logger.error("RunEngine.createRunAttempt(): prisma.$transaction error", {
+                code: error.code,
+                meta: error.meta,
+                stack: error.stack,
+                message: error.message,
+                name: error.name,
+              });
+              throw new ServiceValidationError(
+                "Failed to update task run and execution snapshot",
+                500
+              );
+            }
+          );
+
+          if (!result) {
+            this.logger.error("RunEngine.createRunAttempt(): failed to create task run attempt", {
+              runId: taskRun.id,
+              nextAttemptNumber,
+            });
+            throw new ServiceValidationError("Failed to create task run attempt", 500);
+          }
+
+          const { run, snapshot } = result;
+
+          const machinePreset = getMachinePreset({
+            machines: this.options.machines.machines,
+            defaultMachine: this.options.machines.defaultMachine,
+            config: taskRun.lockedBy.machineConfig ?? {},
+            run: taskRun,
+          });
+
+          const metadata = await parsePacket({
+            data: taskRun.metadata ?? undefined,
+            dataType: taskRun.metadataType,
+          });
+
+          const execution: TaskRunExecution = {
+            task: {
+              id: run.lockedBy!.slug,
+              filePath: run.lockedBy!.filePath,
+              exportName: run.lockedBy!.exportName,
+            },
+            attempt: {
+              number: nextAttemptNumber,
+              startedAt: latestSnapshot.updatedAt,
+              /** @deprecated */
+              id: "deprecated",
+              /** @deprecated */
+              backgroundWorkerId: "deprecated",
+              /** @deprecated */
+              backgroundWorkerTaskId: "deprecated",
+              /** @deprecated */
+              status: "deprecated",
+            },
+            run: {
+              id: run.friendlyId,
+              payload: run.payload,
+              payloadType: run.payloadType,
+              createdAt: run.createdAt,
+              tags: run.tags.map((tag) => tag.name),
+              isTest: run.isTest,
+              idempotencyKey: run.idempotencyKey ?? undefined,
+              startedAt: run.startedAt ?? run.createdAt,
+              maxAttempts: run.maxAttempts ?? undefined,
+              version: run.lockedBy!.worker.version,
+              metadata,
+              maxDuration: run.maxDurationInSeconds ?? undefined,
+              /** @deprecated */
+              context: undefined,
+              /** @deprecated */
+              durationMs: run.usageDurationMs,
+              /** @deprecated */
+              costInCents: run.costInCents,
+              /** @deprecated */
+              baseCostInCents: run.baseCostInCents,
+              traceContext: run.traceContext as Record<string, string | undefined>,
+              priority: run.priorityMs === 0 ? undefined : run.priorityMs / 1_000,
+            },
+            queue: {
+              id: queue.friendlyId,
+              name: queue.name,
+            },
+            environment: {
+              id: environment.id,
+              slug: environment.slug,
+              type: environment.type,
+            },
+            organization: {
+              id: environment.organization.id,
+              slug: environment.organization.slug,
+              name: environment.organization.title,
+            },
+            project: {
+              id: environment.project.id,
+              ref: environment.project.externalRef,
+              slug: environment.project.slug,
+              name: environment.project.name,
+            },
+            batch:
+              taskRun.batchItems[0] && taskRun.batchItems[0].batchTaskRun
+                ? { id: taskRun.batchItems[0].batchTaskRun.friendlyId }
+                : undefined,
+            machine: machinePreset,
+          };
+
+          return { run, snapshot, execution };
         });
-
-        const metadata = await parsePacket({
-          data: taskRun.metadata ?? undefined,
-          dataType: taskRun.metadataType,
-        });
-
-        const execution: TaskRunExecution = {
-          task: {
-            id: run.lockedBy!.slug,
-            filePath: run.lockedBy!.filePath,
-            exportName: run.lockedBy!.exportName,
-          },
-          attempt: {
-            number: nextAttemptNumber,
-            startedAt: latestSnapshot.updatedAt,
-            /** @deprecated */
-            id: "deprecated",
-            /** @deprecated */
-            backgroundWorkerId: "deprecated",
-            /** @deprecated */
-            backgroundWorkerTaskId: "deprecated",
-            /** @deprecated */
-            status: "deprecated",
-          },
-          run: {
-            id: run.friendlyId,
-            payload: run.payload,
-            payloadType: run.payloadType,
-            createdAt: run.createdAt,
-            tags: run.tags.map((tag) => tag.name),
-            isTest: run.isTest,
-            idempotencyKey: run.idempotencyKey ?? undefined,
-            startedAt: run.startedAt ?? run.createdAt,
-            maxAttempts: run.maxAttempts ?? undefined,
-            version: run.lockedBy!.worker.version,
-            metadata,
-            maxDuration: run.maxDurationInSeconds ?? undefined,
-            /** @deprecated */
-            context: undefined,
-            /** @deprecated */
-            durationMs: run.usageDurationMs,
-            /** @deprecated */
-            costInCents: run.costInCents,
-            /** @deprecated */
-            baseCostInCents: run.baseCostInCents,
-            traceContext: run.traceContext as Record<string, string | undefined>,
-            priority: run.priorityMs === 0 ? undefined : run.priorityMs / 1_000,
-          },
-          queue: {
-            id: queue.friendlyId,
-            name: queue.name,
-          },
-          environment: {
-            id: environment.id,
-            slug: environment.slug,
-            type: environment.type,
-          },
-          organization: {
-            id: environment.organization.id,
-            slug: environment.organization.slug,
-            name: environment.organization.title,
-          },
-          project: {
-            id: environment.project.id,
-            ref: environment.project.externalRef,
-            slug: environment.project.slug,
-            name: environment.project.name,
-          },
-          batch:
-            taskRun.batchItems[0] && taskRun.batchItems[0].batchTaskRun
-              ? { id: taskRun.batchItems[0].batchTaskRun.friendlyId }
-              : undefined,
-          machine: machinePreset,
-        };
-
-        return { run, snapshot, execution };
-      });
-    });
+      },
+      {
+        attributes: { runId, snapshotId },
+      }
+    );
   }
 
   /** How a run is completed */
@@ -1412,38 +988,13 @@ export class RunEngine {
     workerId?: string;
     runnerId?: string;
   }): Promise<CompleteRunAttemptResult> {
-    if (completion.metadata) {
-      this.eventBus.emit("runMetadataUpdated", {
-        time: new Date(),
-        run: {
-          id: runId,
-          metadata: completion.metadata,
-        },
-      });
-    }
-
-    switch (completion.ok) {
-      case true: {
-        return this.#attemptSucceeded({
-          runId,
-          snapshotId,
-          completion,
-          tx: this.prisma,
-          workerId,
-          runnerId,
-        });
-      }
-      case false: {
-        return this.#attemptFailed({
-          runId,
-          snapshotId,
-          completion,
-          tx: this.prisma,
-          workerId,
-          runnerId,
-        });
-      }
-    }
+    return this.runAttemptSystem.completeRunAttempt({
+      runId,
+      snapshotId,
+      completion,
+      workerId,
+      runnerId,
+    });
   }
 
   /**
@@ -1469,139 +1020,14 @@ export class RunEngine {
     finalizeRun?: boolean;
     tx?: PrismaClientOrTransaction;
   }): Promise<ExecutionResult> {
-    const prisma = tx ?? this.prisma;
-    reason = reason ?? "Cancelled by user";
-
-    return this.#trace("cancelRun", { runId }, async (span) => {
-      return this.runLock.lock([runId], 5_000, async (signal) => {
-        const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-        //already finished, do nothing
-        if (latestSnapshot.executionStatus === "FINISHED") {
-          return executionResultFromSnapshot(latestSnapshot);
-        }
-
-        //is pending cancellation and we're not finalizing, alert the worker again
-        if (latestSnapshot.executionStatus === "PENDING_CANCEL" && !finalizeRun) {
-          await this.#sendNotificationToWorker({ runId, snapshot: latestSnapshot });
-          return executionResultFromSnapshot(latestSnapshot);
-        }
-
-        //set the run to cancelled immediately
-        const error: TaskRunError = {
-          type: "STRING_ERROR",
-          raw: reason,
-        };
-
-        const run = await prisma.taskRun.update({
-          where: { id: runId },
-          data: {
-            status: "CANCELED",
-            completedAt: finalizeRun ? completedAt ?? new Date() : completedAt,
-            error,
-          },
-          select: {
-            id: true,
-            friendlyId: true,
-            status: true,
-            attemptNumber: true,
-            spanId: true,
-            batchId: true,
-            createdAt: true,
-            completedAt: true,
-            taskEventStore: true,
-            parentTaskRunId: true,
-            runtimeEnvironment: {
-              select: {
-                organizationId: true,
-              },
-            },
-            associatedWaitpoint: {
-              select: {
-                id: true,
-              },
-            },
-            childRuns: {
-              select: {
-                id: true,
-              },
-            },
-          },
-        });
-
-        //remove it from the queue and release concurrency
-        await this.runQueue.acknowledgeMessage(run.runtimeEnvironment.organizationId, runId);
-
-        //if executing, we need to message the worker to cancel the run and put it into `PENDING_CANCEL` status
-        if (isExecuting(latestSnapshot.executionStatus)) {
-          const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-            run,
-            snapshot: {
-              executionStatus: "PENDING_CANCEL",
-              description: "Run was cancelled",
-            },
-            environmentId: latestSnapshot.environmentId,
-            environmentType: latestSnapshot.environmentType,
-            workerId,
-            runnerId,
-          });
-
-          //the worker needs to be notified so it can kill the run and complete the attempt
-          await this.#sendNotificationToWorker({ runId, snapshot: newSnapshot });
-          return executionResultFromSnapshot(newSnapshot);
-        }
-
-        //not executing, so we will actually finish the run
-        const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-          run,
-          snapshot: {
-            executionStatus: "FINISHED",
-            description: "Run was cancelled, not finished",
-          },
-          environmentId: latestSnapshot.environmentId,
-          environmentType: latestSnapshot.environmentType,
-          workerId,
-          runnerId,
-        });
-
-        if (!run.associatedWaitpoint) {
-          throw new ServiceValidationError("No associated waitpoint found", 400);
-        }
-
-        //complete the waitpoint so the parent run can continue
-        await this.completeWaitpoint({
-          id: run.associatedWaitpoint.id,
-          output: { value: JSON.stringify(error), isError: true },
-        });
-
-        this.eventBus.emit("runCancelled", {
-          time: new Date(),
-          run: {
-            id: run.id,
-            friendlyId: run.friendlyId,
-            spanId: run.spanId,
-            taskEventStore: run.taskEventStore,
-            createdAt: run.createdAt,
-            completedAt: run.completedAt,
-            error,
-          },
-        });
-
-        //schedule the cancellation of all the child runs
-        //it will call this function for each child,
-        //which will recursively cancel all children if they need to be
-        if (run.childRuns.length > 0) {
-          for (const childRun of run.childRuns) {
-            await this.worker.enqueue({
-              id: `cancelRun:${childRun.id}`,
-              job: "cancelRun",
-              payload: { runId: childRun.id, completedAt: run.completedAt ?? new Date(), reason },
-            });
-          }
-        }
-
-        return executionResultFromSnapshot(newSnapshot);
-      });
+    return this.runAttemptSystem.cancelRun({
+      runId,
+      workerId,
+      runnerId,
+      completedAt,
+      reason,
+      finalizeRun,
+      tx,
     });
   }
 
@@ -1630,39 +1056,46 @@ export class RunEngine {
     tx?: PrismaClientOrTransaction;
   }): Promise<TaskRun> {
     const prisma = tx ?? this.prisma;
-    return this.#trace("rescheduleRun", { runId }, async (span) => {
-      return await this.runLock.lock([runId], 5_000, async (signal) => {
-        const snapshot = await getLatestExecutionSnapshot(prisma, runId);
+    return startSpan(
+      this.tracer,
+      "rescheduleRun",
+      async (span) => {
+        return await this.runLock.lock([runId], 5_000, async () => {
+          const snapshot = await getLatestExecutionSnapshot(prisma, runId);
 
-        //if the run isn't just created then we can't reschedule it
-        if (snapshot.executionStatus !== "RUN_CREATED") {
-          throw new ServiceValidationError("Cannot reschedule a run that is not delayed");
-        }
+          //if the run isn't just created then we can't reschedule it
+          if (snapshot.executionStatus !== "RUN_CREATED") {
+            throw new ServiceValidationError("Cannot reschedule a run that is not delayed");
+          }
 
-        const updatedRun = await prisma.taskRun.update({
-          where: {
-            id: runId,
-          },
-          data: {
-            delayUntil: delayUntil,
-            executionSnapshots: {
-              create: {
-                engine: "V2",
-                executionStatus: "RUN_CREATED",
-                description: "Delayed run was rescheduled to a future date",
-                runStatus: "EXPIRED",
-                environmentId: snapshot.environmentId,
-                environmentType: snapshot.environmentType,
+          const updatedRun = await prisma.taskRun.update({
+            where: {
+              id: runId,
+            },
+            data: {
+              delayUntil: delayUntil,
+              executionSnapshots: {
+                create: {
+                  engine: "V2",
+                  executionStatus: "RUN_CREATED",
+                  description: "Delayed run was rescheduled to a future date",
+                  runStatus: "EXPIRED",
+                  environmentId: snapshot.environmentId,
+                  environmentType: snapshot.environmentType,
+                },
               },
             },
-          },
+          });
+
+          await this.worker.reschedule(`enqueueDelayedRun:${updatedRun.id}`, delayUntil);
+
+          return updatedRun;
         });
-
-        await this.worker.reschedule(`enqueueDelayedRun:${updatedRun.id}`, delayUntil);
-
-        return updatedRun;
-      });
-    });
+      },
+      {
+        attributes: { runId },
+      }
+    );
   }
 
   async lengthOfEnvQueue(environment: MinimalAuthenticatedEnvironment): Promise<number> {
@@ -1922,8 +1355,6 @@ export class RunEngine {
   async unblockRunForCreatedBatch({
     runId,
     batchId,
-    environmentId,
-    projectId,
     tx,
   }: {
     runId: string;
@@ -1955,20 +1386,12 @@ export class RunEngine {
   }
 
   async tryCompleteBatch({ batchId }: { batchId: string }): Promise<void> {
-    await this.worker.enqueue({
-      //this will debounce the call
-      id: `tryCompleteBatch:${batchId}`,
-      job: "tryCompleteBatch",
-      payload: { batchId: batchId },
-      //2s in the future
-      availableAt: new Date(Date.now() + 2_000),
-    });
+    return this.batchSystem.scheduleCompleteBatch({ batchId });
   }
 
   async getWaitpoint({
     waitpointId,
     environmentId,
-    projectId,
   }: {
     environmentId: string;
     projectId: string;
@@ -2029,7 +1452,7 @@ export class RunEngine {
 
     let $waitpoints = typeof waitpoints === "string" ? [waitpoints] : waitpoints;
 
-    return await this.runLock.lock([runId], 5000, async (signal) => {
+    return await this.runLock.lock([runId], 5000, async () => {
       let snapshot: TaskRunExecutionSnapshot = await getLatestExecutionSnapshot(prisma, runId);
 
       //block the run with the waitpoints, returning how many waitpoints are pending
@@ -2068,7 +1491,7 @@ export class RunEngine {
 
       //if the state has changed, create a new snapshot
       if (newStatus !== snapshot.executionStatus) {
-        snapshot = await this.#createExecutionSnapshot(prisma, {
+        snapshot = await this.executionSnapshotSystem.createExecutionSnapshot(prisma, {
           run: {
             id: snapshot.runId,
             status: snapshot.runStatus,
@@ -2086,7 +1509,7 @@ export class RunEngine {
         });
 
         // Let the worker know immediately, so it can suspend the run
-        await this.#sendNotificationToWorker({ runId, snapshot });
+        await sendNotificationToWorker({ runId, snapshot, eventBus: this.eventBus });
       }
 
       if (timeout) {
@@ -2179,7 +1602,7 @@ export class RunEngine {
     // - Get latest snapshot
     // - If the run is non suspended or going to be, then bail
     // - If the run is suspended or going to be, then release the concurrency
-    await this.runLock.lock([runId], 5_000, async (signal) => {
+    await this.runLock.lock([runId], 5_000, async () => {
       const snapshot = await getLatestExecutionSnapshot(this.prisma, runId);
 
       if (!canReleaseConcurrency(snapshot.executionStatus)) {
@@ -2314,7 +1737,7 @@ export class RunEngine {
   }): Promise<CreateCheckpointResult> {
     const prisma = tx ?? this.prisma;
 
-    return await this.runLock.lock([runId], 5_000, async (signal) => {
+    return await this.runLock.lock([runId], 5_000, async () => {
       const snapshot = await getLatestExecutionSnapshot(prisma, runId);
       if (snapshot.id !== snapshotId) {
         this.eventBus.emit("incomingCheckpointDiscarded", {
@@ -2408,7 +1831,7 @@ export class RunEngine {
       });
 
       //create a new execution snapshot, with the checkpoint
-      const newSnapshot = await this.#createExecutionSnapshot(prisma, {
+      const newSnapshot = await this.executionSnapshotSystem.createExecutionSnapshot(prisma, {
         run,
         snapshot: {
           executionStatus: "SUSPENDED",
@@ -2457,7 +1880,7 @@ export class RunEngine {
   }): Promise<ExecutionResult> {
     const prisma = tx ?? this.prisma;
 
-    return await this.runLock.lock([runId], 5_000, async (signal) => {
+    return await this.runLock.lock([runId], 5_000, async () => {
       const snapshot = await getLatestExecutionSnapshot(prisma, runId);
 
       if (snapshot.id !== snapshotId) {
@@ -2491,7 +1914,7 @@ export class RunEngine {
         throw new ServiceValidationError("Run not found", 404);
       }
 
-      const newSnapshot = await this.#createExecutionSnapshot(prisma, {
+      const newSnapshot = await this.executionSnapshotSystem.createExecutionSnapshot(prisma, {
         run,
         snapshot: {
           executionStatus: "EXECUTING",
@@ -2505,7 +1928,7 @@ export class RunEngine {
       });
 
       // Let worker know about the new snapshot so it can continue the run
-      await this.#sendNotificationToWorker({ runId, snapshot: newSnapshot });
+      await sendNotificationToWorker({ runId, snapshot: newSnapshot, eventBus: this.eventBus });
 
       return {
         ...executionResultFromSnapshot(newSnapshot),
@@ -2532,48 +1955,13 @@ export class RunEngine {
     runnerId?: string;
     tx?: PrismaClientOrTransaction;
   }): Promise<ExecutionResult> {
-    const prisma = tx ?? this.prisma;
-
-    //we don't need to acquire a run lock for any of this, it's not critical if it happens on an older version
-    const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-    if (latestSnapshot.id !== snapshotId) {
-      this.logger.log("heartbeatRun: no longer the latest snapshot, stopping the heartbeat.", {
-        runId,
-        snapshotId,
-        latestSnapshot,
-        workerId,
-        runnerId,
-      });
-
-      await this.worker.ack(`heartbeatSnapshot.${runId}`);
-      return executionResultFromSnapshot(latestSnapshot);
-    }
-
-    if (latestSnapshot.workerId !== workerId) {
-      this.logger.debug("heartbeatRun: worker ID does not match the latest snapshot", {
-        runId,
-        snapshotId,
-        latestSnapshot,
-        workerId,
-        runnerId,
-      });
-    }
-
-    //update the snapshot heartbeat time
-    await prisma.taskRunExecutionSnapshot.update({
-      where: { id: latestSnapshot.id },
-      data: {
-        lastHeartbeatAt: new Date(),
-      },
+    return this.executionSnapshotSystem.heartbeatRun({
+      runId,
+      snapshotId,
+      workerId,
+      runnerId,
+      tx,
     });
-
-    //extending the heartbeat
-    const intervalMs = this.#getHeartbeatIntervalMs(latestSnapshot.executionStatus);
-    if (intervalMs !== null) {
-      await this.worker.reschedule(`heartbeatSnapshot.${runId}`, new Date(Date.now() + intervalMs));
-    }
-
-    return executionResultFromSnapshot(latestSnapshot);
   }
 
   /** Get required data to execute the run */
@@ -2645,52 +2033,9 @@ export class RunEngine {
     }
   }
 
-  async #systemFailure({
-    runId,
-    error,
-    tx,
-  }: {
-    runId: string;
-    error: TaskRunInternalError;
-    tx?: PrismaClientOrTransaction;
-  }): Promise<CompleteRunAttemptResult> {
-    const prisma = tx ?? this.prisma;
-    return this.#trace("#systemFailure", { runId }, async (span) => {
-      const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-      //already finished
-      if (latestSnapshot.executionStatus === "FINISHED") {
-        //todo check run is in the correct state
-        return {
-          attemptStatus: "RUN_FINISHED",
-          snapshot: latestSnapshot,
-          run: {
-            id: runId,
-            friendlyId: latestSnapshot.runFriendlyId,
-            status: latestSnapshot.runStatus,
-            attemptNumber: latestSnapshot.attemptNumber,
-          },
-        };
-      }
-
-      const result = await this.#attemptFailed({
-        runId,
-        snapshotId: latestSnapshot.id,
-        completion: {
-          ok: false,
-          id: runId,
-          error,
-        },
-        tx: prisma,
-      });
-
-      return result;
-    });
-  }
-
   async #expireRun({ runId, tx }: { runId: string; tx?: PrismaClientOrTransaction }) {
     const prisma = tx ?? this.prisma;
-    await this.runLock.lock([runId], 5_000, async (signal) => {
+    await this.runLock.lock([runId], 5_000, async () => {
       const snapshot = await getLatestExecutionSnapshot(prisma, runId);
 
       //if we're executing then we won't expire the run
@@ -2781,515 +2126,7 @@ export class RunEngine {
     });
   }
 
-  async #waitingForDeploy({
-    orgId,
-    runId,
-    workerId,
-    runnerId,
-    reason,
-    tx,
-  }: {
-    orgId: string;
-    runId: string;
-    workerId?: string;
-    runnerId?: string;
-    reason?: string;
-    tx?: PrismaClientOrTransaction;
-  }) {
-    const prisma = tx ?? this.prisma;
-
-    return this.#trace("#waitingForDeploy", { runId }, async (span) => {
-      return this.runLock.lock([runId], 5_000, async (signal) => {
-        //mark run as waiting for deploy
-        const run = await prisma.taskRun.update({
-          where: { id: runId },
-          data: {
-            status: "WAITING_FOR_DEPLOY",
-          },
-          select: {
-            id: true,
-            status: true,
-            attemptNumber: true,
-            runtimeEnvironment: {
-              select: { id: true, type: true },
-            },
-          },
-        });
-
-        await this.#createExecutionSnapshot(prisma, {
-          run,
-          snapshot: {
-            executionStatus: "RUN_CREATED",
-            description:
-              reason ??
-              "The run doesn't have a background worker, so we're going to ack it for now.",
-          },
-          environmentId: run.runtimeEnvironment.id,
-          environmentType: run.runtimeEnvironment.type,
-          workerId,
-          runnerId,
-        });
-
-        //we ack because when it's deployed it will be requeued
-        await this.runQueue.acknowledgeMessage(orgId, runId);
-      });
-    });
-  }
-
-  async #attemptSucceeded({
-    runId,
-    snapshotId,
-    completion,
-    tx,
-    workerId,
-    runnerId,
-  }: {
-    runId: string;
-    snapshotId: string;
-    completion: TaskRunSuccessfulExecutionResult;
-    tx: PrismaClientOrTransaction;
-    workerId?: string;
-    runnerId?: string;
-  }): Promise<CompleteRunAttemptResult> {
-    const prisma = tx ?? this.prisma;
-    return this.#trace("#completeRunAttemptSuccess", { runId, snapshotId }, async (span) => {
-      return this.runLock.lock([runId], 5_000, async (signal) => {
-        const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-        if (latestSnapshot.id !== snapshotId) {
-          throw new ServiceValidationError("Snapshot ID doesn't match the latest snapshot", 400);
-        }
-
-        span.setAttribute("completionStatus", completion.ok);
-
-        const completedAt = new Date();
-
-        const run = await prisma.taskRun.update({
-          where: { id: runId },
-          data: {
-            status: "COMPLETED_SUCCESSFULLY",
-            completedAt,
-            output: completion.output,
-            outputType: completion.outputType,
-            executionSnapshots: {
-              create: {
-                executionStatus: "FINISHED",
-                description: "Task completed successfully",
-                runStatus: "COMPLETED_SUCCESSFULLY",
-                attemptNumber: latestSnapshot.attemptNumber,
-                environmentId: latestSnapshot.environmentId,
-                environmentType: latestSnapshot.environmentType,
-                workerId,
-                runnerId,
-              },
-            },
-          },
-          select: {
-            id: true,
-            friendlyId: true,
-            status: true,
-            attemptNumber: true,
-            spanId: true,
-            associatedWaitpoint: {
-              select: {
-                id: true,
-              },
-            },
-            project: {
-              select: {
-                organizationId: true,
-              },
-            },
-            batchId: true,
-            createdAt: true,
-            completedAt: true,
-            taskEventStore: true,
-            parentTaskRunId: true,
-          },
-        });
-        const newSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-        await this.runQueue.acknowledgeMessage(run.project.organizationId, runId);
-
-        // We need to manually emit this as we created the final snapshot as part of the task run update
-        this.eventBus.emit("executionSnapshotCreated", {
-          time: newSnapshot.createdAt,
-          run: {
-            id: newSnapshot.runId,
-          },
-          snapshot: {
-            ...newSnapshot,
-            completedWaitpointIds: newSnapshot.completedWaitpoints.map((wp) => wp.id),
-          },
-        });
-
-        if (!run.associatedWaitpoint) {
-          throw new ServiceValidationError("No associated waitpoint found", 400);
-        }
-
-        await this.completeWaitpoint({
-          id: run.associatedWaitpoint.id,
-          output: completion.output
-            ? { value: completion.output, type: completion.outputType, isError: false }
-            : undefined,
-        });
-
-        this.eventBus.emit("runSucceeded", {
-          time: completedAt,
-          run: {
-            id: runId,
-            spanId: run.spanId,
-            output: completion.output,
-            outputType: completion.outputType,
-            createdAt: run.createdAt,
-            completedAt: run.completedAt,
-            taskEventStore: run.taskEventStore,
-          },
-        });
-
-        await this.#finalizeRun(run);
-
-        return {
-          attemptStatus: "RUN_FINISHED",
-          snapshot: newSnapshot,
-          run,
-        };
-      });
-    });
-  }
-
-  async #attemptFailed({
-    runId,
-    snapshotId,
-    workerId,
-    runnerId,
-    completion,
-    forceRequeue,
-    tx,
-  }: {
-    runId: string;
-    snapshotId: string;
-    workerId?: string;
-    runnerId?: string;
-    completion: TaskRunFailedExecutionResult;
-    forceRequeue?: boolean;
-    tx: PrismaClientOrTransaction;
-  }): Promise<CompleteRunAttemptResult> {
-    const prisma = this.prisma;
-
-    return this.#trace("completeRunAttemptFailure", { runId, snapshotId }, async (span) => {
-      return this.runLock.lock([runId], 5_000, async (signal) => {
-        const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
-
-        if (latestSnapshot.id !== snapshotId) {
-          throw new ServiceValidationError("Snapshot ID doesn't match the latest snapshot", 400);
-        }
-
-        span.setAttribute("completionStatus", completion.ok);
-
-        //remove waitpoints blocking the run
-        const deletedCount = await this.#clearBlockingWaitpoints({ runId, tx });
-        if (deletedCount > 0) {
-          this.logger.debug("Cleared blocking waitpoints", { runId, deletedCount });
-        }
-
-        const failedAt = new Date();
-
-        const retryResult = await retryOutcomeFromCompletion(prisma, {
-          runId,
-          error: completion.error,
-          retryUsingQueue: forceRequeue ?? false,
-          retrySettings: completion.retry,
-          attemptNumber: latestSnapshot.attemptNumber,
-        });
-
-        // Force requeue means it was crashed so the attempt span needs to be closed
-        if (forceRequeue) {
-          const minimalRun = await prisma.taskRun.findFirst({
-            where: {
-              id: runId,
-            },
-            select: {
-              status: true,
-              spanId: true,
-              maxAttempts: true,
-              runtimeEnvironment: {
-                select: {
-                  organizationId: true,
-                },
-              },
-              taskEventStore: true,
-              createdAt: true,
-              completedAt: true,
-            },
-          });
-
-          if (!minimalRun) {
-            throw new ServiceValidationError("Run not found", 404);
-          }
-
-          this.eventBus.emit("runAttemptFailed", {
-            time: failedAt,
-            run: {
-              id: runId,
-              status: minimalRun.status,
-              spanId: minimalRun.spanId,
-              error: completion.error,
-              attemptNumber: latestSnapshot.attemptNumber ?? 0,
-              createdAt: minimalRun.createdAt,
-              completedAt: minimalRun.completedAt,
-              taskEventStore: minimalRun.taskEventStore,
-            },
-          });
-        }
-
-        switch (retryResult.outcome) {
-          case "cancel_run": {
-            const result = await this.cancelRun({
-              runId,
-              completedAt: failedAt,
-              reason: retryResult.reason,
-              finalizeRun: true,
-              tx: prisma,
-            });
-            return {
-              attemptStatus:
-                result.snapshot.executionStatus === "PENDING_CANCEL"
-                  ? "RUN_PENDING_CANCEL"
-                  : "RUN_FINISHED",
-              ...result,
-            };
-          }
-          case "fail_run": {
-            return await this.#permanentlyFailRun({
-              runId,
-              snapshotId,
-              failedAt,
-              error: retryResult.sanitizedError,
-              workerId,
-              runnerId,
-            });
-          }
-          case "retry": {
-            const retryAt = new Date(retryResult.settings.timestamp);
-
-            const run = await prisma.taskRun.update({
-              where: {
-                id: runId,
-              },
-              data: {
-                status: "RETRYING_AFTER_FAILURE",
-                machinePreset: retryResult.machine,
-              },
-              include: {
-                runtimeEnvironment: {
-                  include: {
-                    project: true,
-                    organization: true,
-                    orgMember: true,
-                  },
-                },
-              },
-            });
-
-            const nextAttemptNumber =
-              latestSnapshot.attemptNumber === null ? 1 : latestSnapshot.attemptNumber + 1;
-
-            if (retryResult.wasOOMError) {
-              this.eventBus.emit("runAttemptFailed", {
-                time: failedAt,
-                run: {
-                  id: runId,
-                  status: run.status,
-                  spanId: run.spanId,
-                  error: completion.error,
-                  attemptNumber: latestSnapshot.attemptNumber ?? 0,
-                  createdAt: run.createdAt,
-                  completedAt: run.completedAt,
-                  taskEventStore: run.taskEventStore,
-                },
-              });
-            }
-
-            this.eventBus.emit("runRetryScheduled", {
-              time: failedAt,
-              run: {
-                id: run.id,
-                friendlyId: run.friendlyId,
-                attemptNumber: nextAttemptNumber,
-                queue: run.queue,
-                taskIdentifier: run.taskIdentifier,
-                traceContext: run.traceContext as Record<string, string | undefined>,
-                baseCostInCents: run.baseCostInCents,
-                spanId: run.spanId,
-              },
-              organization: {
-                id: run.runtimeEnvironment.organizationId,
-              },
-              environment: run.runtimeEnvironment,
-              retryAt,
-            });
-
-            //if it's a long delay and we support checkpointing, put it back in the queue
-            if (
-              forceRequeue ||
-              retryResult.method === "queue" ||
-              (this.options.retryWarmStartThresholdMs !== undefined &&
-                retryResult.settings.delay >= this.options.retryWarmStartThresholdMs)
-            ) {
-              //we nack the message, requeuing it for later
-              const nackResult = await this.#tryNackAndRequeue({
-                run,
-                environment: run.runtimeEnvironment,
-                orgId: run.runtimeEnvironment.organizationId,
-                timestamp: retryAt.getTime(),
-                error: {
-                  type: "INTERNAL_ERROR",
-                  code: "TASK_RUN_DEQUEUED_MAX_RETRIES",
-                  message: `We tried to dequeue the run the maximum number of times but it wouldn't start executing`,
-                },
-                tx: prisma,
-              });
-
-              if (!nackResult.wasRequeued) {
-                return {
-                  attemptStatus: "RUN_FINISHED",
-                  ...nackResult,
-                };
-              } else {
-                return { attemptStatus: "RETRY_QUEUED", ...nackResult };
-              }
-            }
-
-            //it will continue running because the retry delay is short
-            const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-              run,
-              snapshot: {
-                executionStatus: "PENDING_EXECUTING",
-                description: "Attempt failed with a short delay, starting a new attempt",
-              },
-              environmentId: latestSnapshot.environmentId,
-              environmentType: latestSnapshot.environmentType,
-              workerId,
-              runnerId,
-            });
-            //the worker can fetch the latest snapshot and should create a new attempt
-            await this.#sendNotificationToWorker({ runId, snapshot: newSnapshot });
-
-            return {
-              attemptStatus: "RETRY_IMMEDIATELY",
-              ...executionResultFromSnapshot(newSnapshot),
-            };
-          }
-        }
-      });
-    });
-  }
-
-  async #permanentlyFailRun({
-    runId,
-    snapshotId,
-    failedAt,
-    error,
-    workerId,
-    runnerId,
-  }: {
-    runId: string;
-    snapshotId?: string;
-    failedAt: Date;
-    error: TaskRunError;
-    workerId?: string;
-    runnerId?: string;
-  }): Promise<CompleteRunAttemptResult> {
-    const prisma = this.prisma;
-
-    return this.#trace("permanentlyFailRun", { runId, snapshotId }, async (span) => {
-      const status = runStatusFromError(error);
-
-      //run permanently failed
-      const run = await prisma.taskRun.update({
-        where: {
-          id: runId,
-        },
-        data: {
-          status,
-          completedAt: failedAt,
-          error,
-        },
-        select: {
-          id: true,
-          friendlyId: true,
-          status: true,
-          attemptNumber: true,
-          spanId: true,
-          batchId: true,
-          parentTaskRunId: true,
-          associatedWaitpoint: {
-            select: {
-              id: true,
-            },
-          },
-          runtimeEnvironment: {
-            select: {
-              id: true,
-              type: true,
-              organizationId: true,
-            },
-          },
-          taskEventStore: true,
-          createdAt: true,
-          completedAt: true,
-        },
-      });
-
-      const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-        run,
-        snapshot: {
-          executionStatus: "FINISHED",
-          description: "Run failed",
-        },
-        environmentId: run.runtimeEnvironment.id,
-        environmentType: run.runtimeEnvironment.type,
-        workerId,
-        runnerId,
-      });
-
-      if (!run.associatedWaitpoint) {
-        throw new ServiceValidationError("No associated waitpoint found", 400);
-      }
-
-      await this.runQueue.acknowledgeMessage(run.runtimeEnvironment.organizationId, runId);
-
-      await this.completeWaitpoint({
-        id: run.associatedWaitpoint.id,
-        output: { value: JSON.stringify(error), isError: true },
-      });
-
-      this.eventBus.emit("runFailed", {
-        time: failedAt,
-        run: {
-          id: runId,
-          status: run.status,
-          spanId: run.spanId,
-          error,
-          taskEventStore: run.taskEventStore,
-          createdAt: run.createdAt,
-          completedAt: run.completedAt,
-        },
-      });
-
-      await this.#finalizeRun(run);
-
-      return {
-        attemptStatus: "RUN_FINISHED",
-        snapshot: newSnapshot,
-        run,
-      };
-    });
-  }
-
   //MARK: RunQueue
-
   /** The run can be added to the queue. When it's pulled from the queue it will be executed. */
   async #enqueueRun({
     run,
@@ -3322,9 +2159,9 @@ export class RunEngine {
   }): Promise<void> {
     const prisma = tx ?? this.prisma;
 
-    await this.runLock.lock([run.id], 5000, async (signal) => {
-      const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-        run,
+    await this.runLock.lock([run.id], 5000, async () => {
+      const newSnapshot = await this.executionSnapshotSystem.createExecutionSnapshot(prisma, {
+        run: run,
         snapshot: {
           executionStatus: snapshot?.status ?? "QUEUED",
           description: snapshot?.description ?? "Run was QUEUED",
@@ -3359,77 +2196,6 @@ export class RunEngine {
           attempt: 0,
         },
       });
-    });
-  }
-
-  async #tryNackAndRequeue({
-    run,
-    environment,
-    orgId,
-    timestamp,
-    error,
-    workerId,
-    runnerId,
-    tx,
-  }: {
-    run: TaskRun;
-    environment: {
-      id: string;
-      type: RuntimeEnvironmentType;
-    };
-    orgId: string;
-    timestamp?: number;
-    error: TaskRunInternalError;
-    workerId?: string;
-    runnerId?: string;
-    tx?: PrismaClientOrTransaction;
-  }): Promise<{ wasRequeued: boolean } & ExecutionResult> {
-    const prisma = tx ?? this.prisma;
-
-    return await this.runLock.lock([run.id], 5000, async (signal) => {
-      //we nack the message, this allows another work to pick up the run
-      const gotRequeued = await this.runQueue.nackMessage({
-        orgId,
-        messageId: run.id,
-        retryAt: timestamp,
-      });
-
-      if (!gotRequeued) {
-        const result = await this.#systemFailure({
-          runId: run.id,
-          error,
-          tx: prisma,
-        });
-        return { wasRequeued: false, ...result };
-      }
-
-      const newSnapshot = await this.#createExecutionSnapshot(prisma, {
-        run: run,
-        snapshot: {
-          executionStatus: "QUEUED",
-          description: "Requeued the run after a failure",
-        },
-        environmentId: environment.id,
-        environmentType: environment.type,
-        workerId,
-        runnerId,
-      });
-
-      return {
-        wasRequeued: true,
-        snapshot: {
-          id: newSnapshot.id,
-          friendlyId: newSnapshot.friendlyId,
-          executionStatus: newSnapshot.executionStatus,
-          description: newSnapshot.description,
-        },
-        run: {
-          id: newSnapshot.runId,
-          friendlyId: newSnapshot.runFriendlyId,
-          status: newSnapshot.runStatus,
-          attemptNumber: newSnapshot.attemptNumber,
-        },
-      };
     });
   }
 
@@ -3474,7 +2240,7 @@ export class RunEngine {
     }
 
     //4. Continue the run whether it's executing or not
-    await this.runLock.lock([runId], 5000, async (signal) => {
+    await this.runLock.lock([runId], 5000, async () => {
       const snapshot = await getLatestExecutionSnapshot(this.prisma, runId);
 
       //run is still executing, send a message to the worker
@@ -3485,26 +2251,29 @@ export class RunEngine {
         );
 
         if (result) {
-          const newSnapshot = await this.#createExecutionSnapshot(this.prisma, {
-            run: {
-              id: runId,
-              status: snapshot.runStatus,
-              attemptNumber: snapshot.attemptNumber,
-            },
-            snapshot: {
-              executionStatus: "EXECUTING",
-              description: "Run was continued, whilst still executing.",
-            },
-            environmentId: snapshot.environmentId,
-            environmentType: snapshot.environmentType,
-            batchId: snapshot.batchId ?? undefined,
-            completedWaitpoints: blockingWaitpoints.map((b) => ({
-              id: b.waitpoint.id,
-              index: b.batchIndex ?? undefined,
-            })),
-          });
+          const newSnapshot = await this.executionSnapshotSystem.createExecutionSnapshot(
+            this.prisma,
+            {
+              run: {
+                id: runId,
+                status: snapshot.runStatus,
+                attemptNumber: snapshot.attemptNumber,
+              },
+              snapshot: {
+                executionStatus: "EXECUTING",
+                description: "Run was continued, whilst still executing.",
+              },
+              environmentId: snapshot.environmentId,
+              environmentType: snapshot.environmentType,
+              batchId: snapshot.batchId ?? undefined,
+              completedWaitpoints: blockingWaitpoints.map((b) => ({
+                id: b.waitpoint.id,
+                index: b.batchIndex ?? undefined,
+              })),
+            }
+          );
 
-          await this.#sendNotificationToWorker({ runId, snapshot: newSnapshot });
+          await sendNotificationToWorker({ runId, snapshot: newSnapshot, eventBus: this.eventBus });
         } else {
           // Because we cannot reacquire the concurrency, we need to enqueue the run again
           // and because the run is still executing, we need to set the status to QUEUED_EXECUTING
@@ -3698,130 +2467,6 @@ export class RunEngine {
     });
   }
 
-  async #clearBlockingWaitpoints({ runId, tx }: { runId: string; tx?: PrismaClientOrTransaction }) {
-    const prisma = tx ?? this.prisma;
-    const deleted = await prisma.taskRunWaitpoint.deleteMany({
-      where: {
-        taskRunId: runId,
-      },
-    });
-
-    return deleted.count;
-  }
-
-  //#region TaskRunExecutionSnapshots
-  async #createExecutionSnapshot(
-    prisma: PrismaClientOrTransaction,
-    {
-      run,
-      snapshot,
-      batchId,
-      environmentId,
-      environmentType,
-      checkpointId,
-      workerId,
-      runnerId,
-      completedWaitpoints,
-      error,
-    }: {
-      run: { id: string; status: TaskRunStatus; attemptNumber?: number | null };
-      snapshot: {
-        executionStatus: TaskRunExecutionStatus;
-        description: string;
-      };
-      batchId?: string;
-      environmentId: string;
-      environmentType: RuntimeEnvironmentType;
-      checkpointId?: string;
-      workerId?: string;
-      runnerId?: string;
-      completedWaitpoints?: {
-        id: string;
-        index?: number;
-      }[];
-      error?: string;
-    }
-  ) {
-    const newSnapshot = await prisma.taskRunExecutionSnapshot.create({
-      data: {
-        engine: "V2",
-        executionStatus: snapshot.executionStatus,
-        description: snapshot.description,
-        runId: run.id,
-        runStatus: run.status,
-        attemptNumber: run.attemptNumber ?? undefined,
-        batchId,
-        environmentId,
-        environmentType,
-        checkpointId,
-        workerId,
-        runnerId,
-        completedWaitpoints: {
-          connect: completedWaitpoints?.map((w) => ({ id: w.id })),
-        },
-        completedWaitpointOrder: completedWaitpoints
-          ?.filter((c) => c.index !== undefined)
-          .sort((a, b) => a.index! - b.index!)
-          .map((w) => w.id),
-        isValid: error ? false : true,
-        error,
-      },
-      include: {
-        checkpoint: true,
-      },
-    });
-
-    if (!error) {
-      //set heartbeat (if relevant)
-      const intervalMs = this.#getHeartbeatIntervalMs(newSnapshot.executionStatus);
-      if (intervalMs !== null) {
-        await this.worker.enqueue({
-          id: `heartbeatSnapshot.${run.id}`,
-          job: "heartbeatSnapshot",
-          payload: { snapshotId: newSnapshot.id, runId: run.id },
-          availableAt: new Date(Date.now() + intervalMs),
-        });
-      }
-    }
-
-    this.eventBus.emit("executionSnapshotCreated", {
-      time: newSnapshot.createdAt,
-      run: {
-        id: newSnapshot.runId,
-      },
-      snapshot: {
-        ...newSnapshot,
-        completedWaitpointIds: completedWaitpoints?.map((w) => w.id) ?? [],
-      },
-    });
-
-    return {
-      ...newSnapshot,
-      friendlyId: SnapshotId.toFriendlyId(newSnapshot.id),
-      runFriendlyId: RunId.toFriendlyId(newSnapshot.runId),
-    };
-  }
-
-  #getHeartbeatIntervalMs(status: TaskRunExecutionStatus): number | null {
-    switch (status) {
-      case "PENDING_EXECUTING": {
-        return this.heartbeatTimeouts.PENDING_EXECUTING;
-      }
-      case "PENDING_CANCEL": {
-        return this.heartbeatTimeouts.PENDING_CANCEL;
-      }
-      case "EXECUTING": {
-        return this.heartbeatTimeouts.EXECUTING;
-      }
-      case "EXECUTING_WITH_WAITPOINTS": {
-        return this.heartbeatTimeouts.EXECUTING_WITH_WAITPOINTS;
-      }
-      default: {
-        return null;
-      }
-    }
-  }
-
   //#endregion
 
   //#region Heartbeat
@@ -3835,7 +2480,7 @@ export class RunEngine {
     tx?: PrismaClientOrTransaction;
   }) {
     const prisma = tx ?? this.prisma;
-    return await this.runLock.lock([runId], 5_000, async (signal) => {
+    return await this.runLock.lock([runId], 5_000, async () => {
       const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
       if (latestSnapshot.id !== snapshotId) {
         this.logger.log(
@@ -3909,20 +2554,6 @@ export class RunEngine {
           }
 
           //it will automatically be requeued X times depending on the queue retry settings
-          const gotRequeued = await this.#tryNackAndRequeue({
-            run,
-            environment: {
-              id: latestSnapshot.environmentId,
-              type: latestSnapshot.environmentType,
-            },
-            orgId: run.runtimeEnvironment.organizationId,
-            error: {
-              type: "INTERNAL_ERROR",
-              code: "TASK_RUN_DEQUEUED_MAX_RETRIES",
-              message: `Trying to create an attempt failed multiple times, exceeding how many times we retry.`,
-            },
-            tx: prisma,
-          });
           break;
         }
         case "EXECUTING":
@@ -3930,7 +2561,7 @@ export class RunEngine {
           const retryDelay = 250;
 
           //todo call attemptFailed and force requeuing
-          await this.#attemptFailed({
+          await this.runAttemptSystem.attemptFailed({
             runId,
             snapshotId: latestSnapshot.id,
             completion: {
@@ -3982,103 +2613,6 @@ export class RunEngine {
 
   //#endregion
 
-  /**
-   * Sends a notification that a run has changed and we need to fetch the latest run state.
-   * The worker will call `getRunExecutionData` via the API and act accordingly.
-   */
-  async #sendNotificationToWorker({
-    runId,
-    snapshot,
-  }: {
-    runId: string;
-    snapshot: {
-      id: string;
-      executionStatus: TaskRunExecutionStatus;
-    };
-  }) {
-    this.eventBus.emit("workerNotification", {
-      time: new Date(),
-      run: {
-        id: runId,
-      },
-      snapshot: {
-        id: snapshot.id,
-        executionStatus: snapshot.executionStatus,
-      },
-    });
-  }
-
-  /*
-   * Whether the run succeeds, fails, is cancelled… we need to run these operations
-   */
-  async #finalizeRun({ id, batchId }: { id: string; batchId: string | null }) {
-    if (batchId) {
-      await this.tryCompleteBatch({ batchId });
-    }
-
-    //cancel the heartbeats
-    await this.worker.ack(`heartbeatSnapshot.${id}`);
-  }
-
-  /**
-   * Checks to see if all runs for a BatchTaskRun are completed, if they are then update the status.
-   * This isn't used operationally, but it's used for the Batches dashboard page.
-   */
-  async #tryCompleteBatch({ batchId }: { batchId: string }) {
-    return this.#trace(
-      "#tryCompleteBatch",
-      {
-        batchId,
-      },
-      async (span) => {
-        const batch = await this.prisma.batchTaskRun.findUnique({
-          select: {
-            status: true,
-            runtimeEnvironmentId: true,
-          },
-          where: {
-            id: batchId,
-          },
-        });
-
-        if (!batch) {
-          this.logger.error("#tryCompleteBatch batch doesn't exist", { batchId });
-          return;
-        }
-
-        if (batch.status === "COMPLETED") {
-          this.logger.debug("#tryCompleteBatch: Batch already completed", { batchId });
-          return;
-        }
-
-        const runs = await this.prisma.taskRun.findMany({
-          select: {
-            id: true,
-            status: true,
-          },
-          where: {
-            batchId,
-            runtimeEnvironmentId: batch.runtimeEnvironmentId,
-          },
-        });
-
-        if (runs.every((r) => isFinalRunStatus(r.status))) {
-          this.logger.debug("#tryCompleteBatch: All runs are completed", { batchId });
-          await this.prisma.batchTaskRun.update({
-            where: {
-              id: batchId,
-            },
-            data: {
-              status: "COMPLETED",
-            },
-          });
-        } else {
-          this.logger.debug("#tryCompleteBatch: Not all runs are completed", { batchId });
-        }
-      }
-    );
-  }
-
   async #getAuthenticatedEnvironmentFromRun(runId: string, tx?: PrismaClientOrTransaction) {
     const prisma = tx ?? this.prisma;
     const taskRun = await prisma.taskRun.findUnique({
@@ -4108,36 +2642,6 @@ export class RunEngine {
 
   #backgroundWorkerQueueKey(backgroundWorkerId: string) {
     return `master-background-worker:${backgroundWorkerId}`;
-  }
-
-  async #trace<T>(
-    trace: string,
-    attributes: Attributes | undefined,
-    fn: (span: Span) => Promise<T>
-  ): Promise<T> {
-    return this.tracer.startActiveSpan(
-      `${this.constructor.name}.${trace}`,
-      { attributes, kind: SpanKind.SERVER },
-      async (span) => {
-        try {
-          return await fn(span);
-        } catch (e) {
-          if (e instanceof ServiceValidationError) {
-            throw e;
-          }
-
-          if (e instanceof Error) {
-            span.recordException(e);
-          } else {
-            span.recordException(new Error(String(e)));
-          }
-
-          throw e;
-        } finally {
-          span.end();
-        }
-      }
-    );
   }
 }
