+++ internal-packages/run-engine/src/engine/systems/executionSnapshotSystem.ts
@@ -1,0 +1,335 @@
+import {
+  PrismaClient,
+  PrismaClientOrTransaction,
+  RuntimeEnvironmentType,
+  TaskRunExecutionStatus,
+  TaskRunStatus,
+  TaskRunCheckpoint,
+  TaskRunExecutionSnapshot,
+} from "@trigger.dev/database";
+import { EngineWorker, HeartbeatTimeouts } from "../types.js";
+import { EventBus } from "../eventBus.js";
+import { CompletedWaitpoint, ExecutionResult } from "@trigger.dev/core/v3";
+import { BatchId, RunId, SnapshotId } from "@trigger.dev/core/v3/isomorphic";
+import { Logger } from "@trigger.dev/core/logger";
+import { Tracer } from "@internal/tracing";
+
+export type ExecutionSnapshotSystemOptions = {
+  prisma: PrismaClient;
+  logger: Logger;
+  tracer: Tracer;
+  worker: EngineWorker;
+  eventBus: EventBus;
+  heartbeatTimeouts: HeartbeatTimeouts;
+};
+
+export interface LatestExecutionSnapshot extends TaskRunExecutionSnapshot {
+  friendlyId: string;
+  runFriendlyId: string;
+  checkpoint: TaskRunCheckpoint | null;
+  completedWaitpoints: CompletedWaitpoint[];
+}
+
+/* Gets the most recent valid snapshot for a run */
+export async function getLatestExecutionSnapshot(
+  prisma: PrismaClientOrTransaction,
+  runId: string
+): Promise<LatestExecutionSnapshot> {
+  const snapshot = await prisma.taskRunExecutionSnapshot.findFirst({
+    where: { runId, isValid: true },
+    include: {
+      completedWaitpoints: true,
+      checkpoint: true,
+    },
+    orderBy: { createdAt: "desc" },
+  });
+
+  if (!snapshot) {
+    throw new Error(`No execution snapshot found for TaskRun ${runId}`);
+  }
+
+  return {
+    ...snapshot,
+    friendlyId: SnapshotId.toFriendlyId(snapshot.id),
+    runFriendlyId: RunId.toFriendlyId(snapshot.runId),
+    completedWaitpoints: snapshot.completedWaitpoints.flatMap((w) => {
+      //get all indexes of the waitpoint in the completedWaitpointOrder
+      //we do this because the same run can be in a batch multiple times (i.e. same idempotencyKey)
+      let indexes: (number | undefined)[] = [];
+      for (let i = 0; i < snapshot.completedWaitpointOrder.length; i++) {
+        if (snapshot.completedWaitpointOrder[i] === w.id) {
+          indexes.push(i);
+        }
+      }
+
+      if (indexes.length === 0) {
+        indexes.push(undefined);
+      }
+
+      return indexes.map((index) => {
+        return {
+          id: w.id,
+          index: index === -1 ? undefined : index,
+          friendlyId: w.friendlyId,
+          type: w.type,
+          completedAt: w.completedAt ?? new Date(),
+          idempotencyKey:
+            w.userProvidedIdempotencyKey && !w.inactiveIdempotencyKey
+              ? w.idempotencyKey
+              : undefined,
+          completedByTaskRun: w.completedByTaskRunId
+            ? {
+                id: w.completedByTaskRunId,
+                friendlyId: RunId.toFriendlyId(w.completedByTaskRunId),
+                batch: snapshot.batchId
+                  ? {
+                      id: snapshot.batchId,
+                      friendlyId: BatchId.toFriendlyId(snapshot.batchId),
+                    }
+                  : undefined,
+              }
+            : undefined,
+          completedAfter: w.completedAfter ?? undefined,
+          completedByBatch: w.completedByBatchId
+            ? {
+                id: w.completedByBatchId,
+                friendlyId: BatchId.toFriendlyId(w.completedByBatchId),
+              }
+            : undefined,
+          output: w.output ?? undefined,
+          outputType: w.outputType,
+          outputIsError: w.outputIsError,
+        } satisfies CompletedWaitpoint;
+      });
+    }),
+  };
+}
+
+export async function getExecutionSnapshotCompletedWaitpoints(
+  prisma: PrismaClientOrTransaction,
+  snapshotId: string
+) {
+  const waitpoints = await prisma.taskRunExecutionSnapshot.findFirst({
+    where: { id: snapshotId },
+    include: {
+      completedWaitpoints: true,
+    },
+  });
+
+  //deduplicate waitpoints
+  const waitpointIds = new Set<string>();
+  return (
+    waitpoints?.completedWaitpoints.filter((waitpoint) => {
+      if (waitpointIds.has(waitpoint.id)) {
+        return false;
+      } else {
+        waitpointIds.add(waitpoint.id);
+        return true;
+      }
+    }) ?? []
+  );
+}
+
+export function executionResultFromSnapshot(snapshot: TaskRunExecutionSnapshot): ExecutionResult {
+  return {
+    snapshot: {
+      id: snapshot.id,
+      friendlyId: SnapshotId.toFriendlyId(snapshot.id),
+      executionStatus: snapshot.executionStatus,
+      description: snapshot.description,
+    },
+    run: {
+      id: snapshot.runId,
+      friendlyId: RunId.toFriendlyId(snapshot.runId),
+      status: snapshot.runStatus,
+      attemptNumber: snapshot.attemptNumber,
+    },
+  };
+}
+
+export class ExecutionSnapshotSystem {
+  private readonly worker: EngineWorker;
+  private readonly eventBus: EventBus;
+  private readonly heartbeatTimeouts: HeartbeatTimeouts;
+  private readonly prisma: PrismaClient;
+  private readonly logger: Logger;
+  private readonly tracer: Tracer;
+
+  constructor(private readonly options: ExecutionSnapshotSystemOptions) {
+    this.worker = options.worker;
+    this.eventBus = options.eventBus;
+    this.heartbeatTimeouts = options.heartbeatTimeouts;
+    this.prisma = options.prisma;
+    this.logger = options.logger;
+    this.tracer = options.tracer;
+  }
+
+  public async createExecutionSnapshot(
+    prisma: PrismaClientOrTransaction,
+    {
+      run,
+      snapshot,
+      batchId,
+      environmentId,
+      environmentType,
+      checkpointId,
+      workerId,
+      runnerId,
+      completedWaitpoints,
+      error,
+    }: {
+      run: { id: string; status: TaskRunStatus; attemptNumber?: number | null };
+      snapshot: {
+        executionStatus: TaskRunExecutionStatus;
+        description: string;
+      };
+      batchId?: string;
+      environmentId: string;
+      environmentType: RuntimeEnvironmentType;
+      checkpointId?: string;
+      workerId?: string;
+      runnerId?: string;
+      completedWaitpoints?: {
+        id: string;
+        index?: number;
+      }[];
+      error?: string;
+    }
+  ) {
+    const newSnapshot = await prisma.taskRunExecutionSnapshot.create({
+      data: {
+        engine: "V2",
+        executionStatus: snapshot.executionStatus,
+        description: snapshot.description,
+        runId: run.id,
+        runStatus: run.status,
+        attemptNumber: run.attemptNumber ?? undefined,
+        batchId,
+        environmentId,
+        environmentType,
+        checkpointId,
+        workerId,
+        runnerId,
+        completedWaitpoints: {
+          connect: completedWaitpoints?.map((w) => ({ id: w.id })),
+        },
+        completedWaitpointOrder: completedWaitpoints
+          ?.filter((c) => c.index !== undefined)
+          .sort((a, b) => a.index! - b.index!)
+          .map((w) => w.id),
+        isValid: error ? false : true,
+        error,
+      },
+      include: {
+        checkpoint: true,
+      },
+    });
+
+    if (!error) {
+      //set heartbeat (if relevant)
+      const intervalMs = this.#getHeartbeatIntervalMs(newSnapshot.executionStatus);
+      if (intervalMs !== null) {
+        await this.worker.enqueue({
+          id: `heartbeatSnapshot.${run.id}`,
+          job: "heartbeatSnapshot",
+          payload: { snapshotId: newSnapshot.id, runId: run.id },
+          availableAt: new Date(Date.now() + intervalMs),
+        });
+      }
+    }
+
+    this.eventBus.emit("executionSnapshotCreated", {
+      time: newSnapshot.createdAt,
+      run: {
+        id: newSnapshot.runId,
+      },
+      snapshot: {
+        ...newSnapshot,
+        completedWaitpointIds: completedWaitpoints?.map((w) => w.id) ?? [],
+      },
+    });
+
+    return {
+      ...newSnapshot,
+      friendlyId: SnapshotId.toFriendlyId(newSnapshot.id),
+      runFriendlyId: RunId.toFriendlyId(newSnapshot.runId),
+    };
+  }
+
+  public async heartbeatRun({
+    runId,
+    snapshotId,
+    workerId,
+    runnerId,
+    tx,
+  }: {
+    runId: string;
+    snapshotId: string;
+    workerId?: string;
+    runnerId?: string;
+    tx?: PrismaClientOrTransaction;
+  }): Promise<ExecutionResult> {
+    const prisma = tx ?? this.prisma;
+
+    //we don't need to acquire a run lock for any of this, it's not critical if it happens on an older version
+    const latestSnapshot = await getLatestExecutionSnapshot(prisma, runId);
+    if (latestSnapshot.id !== snapshotId) {
+      this.logger.log("heartbeatRun: no longer the latest snapshot, stopping the heartbeat.", {
+        runId,
+        snapshotId,
+        latestSnapshot,
+        workerId,
+        runnerId,
+      });
+
+      await this.worker.ack(`heartbeatSnapshot.${runId}`);
+      return executionResultFromSnapshot(latestSnapshot);
+    }
+
+    if (latestSnapshot.workerId !== workerId) {
+      this.logger.debug("heartbeatRun: worker ID does not match the latest snapshot", {
+        runId,
+        snapshotId,
+        latestSnapshot,
+        workerId,
+        runnerId,
+      });
+    }
+
+    //update the snapshot heartbeat time
+    await prisma.taskRunExecutionSnapshot.update({
+      where: { id: latestSnapshot.id },
+      data: {
+        lastHeartbeatAt: new Date(),
+      },
+    });
+
+    //extending the heartbeat
+    const intervalMs = this.#getHeartbeatIntervalMs(latestSnapshot.executionStatus);
+    if (intervalMs !== null) {
+      await this.worker.reschedule(`heartbeatSnapshot.${runId}`, new Date(Date.now() + intervalMs));
+    }
+
+    return executionResultFromSnapshot(latestSnapshot);
+  }
+
+  #getHeartbeatIntervalMs(status: TaskRunExecutionStatus): number | null {
+    switch (status) {
+      case "PENDING_EXECUTING": {
+        return this.heartbeatTimeouts.PENDING_EXECUTING;
+      }
+      case "PENDING_CANCEL": {
+        return this.heartbeatTimeouts.PENDING_CANCEL;
+      }
+      case "EXECUTING": {
+        return this.heartbeatTimeouts.EXECUTING;
+      }
+      case "EXECUTING_WITH_WAITPOINTS": {
+        return this.heartbeatTimeouts.EXECUTING_WITH_WAITPOINTS;
+      }
+      default: {
+        return null;
+      }
+    }
+  }
+}
