+++ internal-packages/run-engine/src/engine/workerCatalog.ts
@@ -1,0 +1,56 @@
+import { z } from "zod";
+
+export const workerCatalog = {
+  finishWaitpoint: {
+    schema: z.object({
+      waitpointId: z.string(),
+      error: z.string().optional(),
+    }),
+    visibilityTimeoutMs: 5000,
+  },
+  heartbeatSnapshot: {
+    schema: z.object({
+      runId: z.string(),
+      snapshotId: z.string(),
+    }),
+    visibilityTimeoutMs: 5000,
+  },
+  expireRun: {
+    schema: z.object({
+      runId: z.string(),
+    }),
+    visibilityTimeoutMs: 5000,
+  },
+  cancelRun: {
+    schema: z.object({
+      runId: z.string(),
+      completedAt: z.coerce.date(),
+      reason: z.string().optional(),
+    }),
+    visibilityTimeoutMs: 5000,
+  },
+  queueRunsWaitingForWorker: {
+    schema: z.object({
+      backgroundWorkerId: z.string(),
+    }),
+    visibilityTimeoutMs: 5000,
+  },
+  tryCompleteBatch: {
+    schema: z.object({
+      batchId: z.string(),
+    }),
+    visibilityTimeoutMs: 10_000,
+  },
+  continueRunIfUnblocked: {
+    schema: z.object({
+      runId: z.string(),
+    }),
+    visibilityTimeoutMs: 10_000,
+  },
+  enqueueDelayedRun: {
+    schema: z.object({
+      runId: z.string(),
+    }),
+    visibilityTimeoutMs: 10_000,
+  },
+};
